requirements.txt
discord_webhook==1.1.0
httpx==0.23.0
python-dotenv==1.0.0
rich==13.3.2
